+++ src/domd/__init__.py
@@ -1,4 +1,4 @@
-"""TodoMD Package"""
+"""DoMD Package"""
 
 __version__ = "0.1.0"